requirements.txt
alembic==1.3.3
Babel==2.8.0
bcrypt==3.1.7
blinker==1.4
boto3==1.12.25
botocore==1.15.25
cachetools==4.0.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
decorator==4.4.2
docutils==0.15.2
dominate==2.4.0
Flask==1.1.1
Flask-Admin==1.5.5
Flask-Babel==1.0.0
Flask-Babel2==0.1
Flask-BabelEx==0.9.4
Flask-Bcrypt==0.7.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-WhooshAlchemyPlus==0.7.4
flask-whooshee==0.7.0
Flask-WTF==0.14.2
gensim==3.8.1
google-api-core==1.16.0
google-auth==1.11.3
google-cloud-core==1.3.0
google-cloud-storage==1.26.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
idna==2.9
itsdangerous==1.1.0
jieba==0.42.1
Jinja2==2.11.1
jmespath==0.9.5
Mako==1.1.1
MarkupSafe==1.1.1
numpy==1.18.2
Pillow==7.0.0
protobuf==3.11.3
public==2019.4.13
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
PyJWT==1.7.1
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2019.3
requests==2.23.0
rsa==4.0
s3transfer==0.3.3
scipy==1.4.1
self==2019.4.13
six==1.14.0
smart-open==1.10.0
speaklater==1.3
SQLAlchemy==1.3.13
urllib3==1.25.8
visitor==0.1.3
Werkzeug==0.16.1
Whoosh==2.7.4
WTForms==2.2.1
